trainer/requirements.txt
pymongo
scikit-learn
joblib
pandas
python-dotenv
tensorflow
numpy